Improve history of reservation on destruction.
There was a problem before where it the history of a destroyed record
was the destroyed record rather than what it was before it was
destroyed. This adds a new hook so it displays appropriately.

This also renames "Reserved At" in the history view to "Updated At" to
more appropriately reflect what it does.

Closes #253